Moon_Dance/Moon_Dance/requirements.txt
numpy>=1.21.0
pandas
tk
requests
fastapi
uvicorn
pydantic
Flask
Flask-Cors
PyJWT
pyOpenSSL
celery>=5.3.0
redis>=5.0.0
psutil>=5.9.0
